对回答问题页面进行修改，删除无用的静态习题代码块，以及静态的题目序号显示，添加计时器功能，添加主页index及相关的静态资源

+++ src/main/java/com/pojo/TbClassify.java
@@ -3,15 +3,15 @@ package com.pojo;
 
 public class TbClassify {
 
-  private long classifyId;
+  private Integer classifyId;
   private String classifyName;
 
 
-  public long getClassifyId() {
+  public Integer getClassifyId() {
     return classifyId;
   }
 
-  public void setClassifyId(long classifyId) {
+  public void setClassifyId(Integer classifyId) {
     this.classifyId = classifyId;
   }
